// src/Interfaces/ITicketR.ts
export interface ITicketR {
  codigoTicket: number;
  tituloTicket: string;
  descripcionTicket: string;
  nombreCliente: string;
  nombreCategoria: string;
  nombrePrioridad: string;
  nombreDepto: string;
  nombreUsuario: string;
  nombreEstado: string;
  fechaCreacion: string;
  fechaModificacion: string;
  fechaCierre: string;
}